+++ src/dynamic_boundary_conditions/rainfall/main_rainfall.py
@@ -23,21 +23,6 @@ from src.dynamic_boundary_conditions.rainfall import (
 )
 
 
-def remove_existing_rain_inputs(bg_flood_dir: pathlib.Path) -> None:
-    """
-    Remove existing rain input files from the specified directory.
-
-    Parameters
-    ----------
-    bg_flood_dir : pathlib.Path
-        BG-Flood model directory containing the rain input files.
-    """
-    # Iterate through all rain input files in the directory
-    for rain_input_file in bg_flood_dir.glob('rain_forcing.*'):
-        # Remove the file
-        rain_input_file.unlink()
-
-
 def main(
         selected_polygon_gdf: gpd.GeoDataFrame,
         rcp: Optional[float],
@@ -94,7 +79,7 @@ def main(
     # BG-Flood Model Directory
     bg_flood_dir = config.get_env_variable("FLOOD_MODEL_DIR", cast_to=pathlib.Path)
     # Remove any existing rainfall model inputs in the BG-Flood directory
-    remove_existing_rain_inputs(bg_flood_dir)
+    rainfall_model_input.remove_existing_rain_inputs(bg_flood_dir)
 
     # Fetch rainfall sites data from the HIRDS website and store it to the database
     rainfall_sites.rainfall_sites_to_db(engine)
